A: Refactored ScreenNavigator to use simple-stack, Renamed noteslist package
