+++ android/src/main/java/com/akjaw/fullerstack/dependencyinjection/modules/presentationModule.kt
@@ -1,7 +1,7 @@
 package com.akjaw.fullerstack.dependencyinjection.modules
 
 import com.akjaw.fullerstack.screens.common.ViewMvcFactory
-import com.akjaw.fullerstack.screens.noteslist.NotesListController
+import com.akjaw.fullerstack.screens.list.NotesListController
 import org.kodein.di.DI
 import org.kodein.di.bind
 import org.kodein.di.instance
