--- android/src/main/java/com/akjaw/fullerstack/screens/common/ScreenNavigator.kt
@@ -1,7 +1,0 @@
-package com.akjaw.fullerstack.screens.common
-
-interface ScreenNavigator {
-
-    fun openAddNoteScreen()
-
-}
