+++ android/src/main/java/com/akjaw/fullerstack/screens/common/navigation/ScreenNavigator.kt
@@ -1,0 +1,11 @@
+package com.akjaw.fullerstack.screens.common.navigation
+
+import android.content.Context
+
+interface ScreenNavigator {
+
+    fun openAddNoteScreen(context: Context)
+
+    fun goBack(context: Context)
+
+}
