+++ android/src/main/java/com/akjaw/fullerstack/screens/common/navigation/keys/NoteEditorScreen.kt
@@ -1,4 +1,4 @@
-package com.akjaw.fullerstack.screens.common.navigation
+package com.akjaw.fullerstack.screens.common.navigation.keys
 
 import androidx.fragment.app.Fragment
 import com.akjaw.fullerstack.screens.editor.NoteEditorFragment
