+++ android/src/main/java/com/akjaw/fullerstack/screens/common/navigation/keys/NotesListScreen.kt
@@ -1,7 +1,7 @@
-package com.akjaw.fullerstack.screens.common.navigation
+package com.akjaw.fullerstack.screens.common.navigation.keys
 
 import androidx.fragment.app.Fragment
-import com.akjaw.fullerstack.screens.noteslist.NotesListFragment
+import com.akjaw.fullerstack.screens.list.NotesListFragment
 import com.zhuinden.simplestackextensions.fragments.DefaultFragmentKey
 import kotlinx.android.parcel.Parcelize
