+++ android/src/main/java/com/akjaw/fullerstack/screens/list/NotesListViewMvc.kt
@@ -1,4 +1,4 @@
-package com.akjaw.fullerstack.screens.noteslist
+package com.akjaw.fullerstack.screens.list
 
 import com.akjaw.fullerstack.screens.common.base.BaseObservableViewMvc
 import data.Note
